+++ src/main/java/org/scify/jedai/progressivejoin/JaccardTopK.java
@@ -63,8 +63,8 @@ public class JaccardTopK {
         return len1 - minsize(len1, thres) + 1;
     }
 
-    public double computesim(int len1, int len2, int overlap) {
-        return overlap / (0.0 + len1 + len2 - overlap);
+    public float computesim(int len1, int len2, int overlap) {
+        return (float)(overlap / (0.0 + len1 + len2 - overlap));
     }
 
     public double upperbound_access(int len1, int len2, int pos1, int pos2) {
