+++ src/main/java/org/scify/jedai/similarityjoins/characterbased/AllPairs.java
@@ -252,7 +252,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                 int distance = getEditDistance(attributeValues.get(x), attributeValues.get(y), threshold);
                 if (distance <= threshold) {
                     final Comparison currentComp = getComparison(originalId[x], originalId[y]);
-                    currentComp.setUtilityMeasure(1 - (double) distance / threshold);
+                    currentComp.setUtilityMeasure(1 - (float) distance / threshold);
                     executedComparisons.add(currentComp);
                 }
             }
@@ -314,7 +314,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                     continue;
                 }
 
-                double distance = getEditDistance(attributeValues.get(k), attributeValues.get(cand), threshold);
+                float distance = getEditDistance(attributeValues.get(k), attributeValues.get(cand), threshold);
                 if (distance <= threshold) {
                     int id1 = Math.min(originalId[k], originalId[cand]);
                     int id2 = Math.max(originalId[k], originalId[cand]);
@@ -353,7 +353,7 @@ public class AllPairs extends AbstractCharacterBasedJoin {
                         continue;
                     }
                 }
-                double distance = getEditDistance(attributeValues.get(k), attributeValues.get(bound), threshold);
+                float distance = getEditDistance(attributeValues.get(k), attributeValues.get(bound), threshold);
                 if (distance <= threshold) {
                     int id1 = Math.min(originalId[k], originalId[bound]);
                     int id2 = Math.max(originalId[k], originalId[bound]);
